RP-366 Create Sandbox API/Service Moved the AggregateReporting config file to the common configuration package. Updated files
