+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/ApplicationSandboxConfiguration.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.admin.model;
 
 import org.opentestsystem.rdw.multitenant.TenantProperties;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemPropertiesImpl;
 
 import java.util.Set;
