+++ admin-service/src/main/java/org/opentestsystem/rdw/admin/model/SandboxConfigurationStubs.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.admin.model;
 
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 
 import java.util.HashSet;
 import java.util.LinkedHashMap;
