+++ admin-service/src/test/java/org/opentestsystem/rdw/admin/service/impl/DefaultSandboxServiceTest.java
@@ -8,7 +8,7 @@ import org.opentestsystem.rdw.admin.model.ApplicationSandboxConfigurations;
 import org.opentestsystem.rdw.admin.model.SandboxConfiguration;
 import org.opentestsystem.rdw.admin.model.SandboxConfigurationStubs;
 import org.opentestsystem.rdw.admin.repository.impl.InMemorySandboxConfigurationRepository;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.reporting.common.configuration.ReportingSystemProperties;
 
 import java.util.Map;
