+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/AbstractAggregateQueryHandler.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.olap.service.impl;
 
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
 import org.opentestsystem.rdw.olap.model.AggregateReportQueryDescription;
 import org.opentestsystem.rdw.olap.model.AggregateRowReportResult;
 import org.opentestsystem.rdw.olap.model.OrganizationTypeQuery;
