+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/configuration/AggregateReportingPropertiesResolverIT.java
@@ -3,6 +3,7 @@ package org.opentestsystem.rdw.olap.configuration;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opentestsystem.rdw.multitenant.TenantContextHolder;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesResolver;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
