+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/DefaultUserAssessmentTypeServiceTest.java
@@ -4,7 +4,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
 import org.junit.Before;
 import org.junit.Test;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.olap.service.AssessmentTypeService;
 import org.opentestsystem.rdw.reporting.common.model.CodedEntity;
 import org.opentestsystem.rdw.reporting.common.security.User;
