+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/TargetReportQueryHandlerTest.java
@@ -5,8 +5,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
 import org.mockito.ArgumentCaptor;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingProperties;
-import org.opentestsystem.rdw.olap.configuration.AggregateReportingPropertiesTenant;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingProperties;
+import org.opentestsystem.rdw.reporting.common.configuration.AggregateReportingPropertiesTenant;
 import org.opentestsystem.rdw.olap.model.AggregateReportQueryDescription;
 import org.opentestsystem.rdw.olap.model.SubjectAssessment;
 import org.opentestsystem.rdw.olap.model.TargetOrganizationQuery;
